FEAT: Add a new component Hero Editor
The application now has a basic title. Next you will create a new component to display hero information and place that component in the application shell.

Used the CLI to create a second Heroescomponent.
Displayed the HeroesComponent by adding it to the Appcomponent shell.
Applied the UppercasePie to format the name.
Used ttwo-way data binding with the ngModel directive.
Learned about the appModule
Imported the formsModule in the AppModule so that Angula would recognize and apply the ngModel directive.
Learned the importance of declaring components in the AppModule and appreciated that the CLI declared it directly.